Added classes and interfaces to partition given program w.r.t. annotations
ProgramPartitionAnalysis - analysis pass to find functions that should be partitioned and moved to an enclave (or
library).
ProgramPartition - the partition for given annotations. Partition is a set of functions
Partitioner - runner class that does the actual detection of functions to be put inside a partition.
Different implementations of Partitioner are responsible to do the job for different types of Annotations, i.e.
partition for function only, for output arguments, for return value.

Added check of annotation name equality in Annotation operator ==
Added default constructor for Annotation

Added an interface to AnnotationParser to request all annotations
Added UML of the project as an image and xml.