Static analyser assigns level of relations between functions and security sensitive functions
Instead of adding functions found during static analysis to the set of security sensitive functions assign level of relations for them and the user annotated functions.
The level represent how closely is a function found by static analyser related to a security sensitive functions.
If a function has different levels for different security sensitive functions the minimum is taken